Service for getFinalizedTestResults
